updated all database SQL-Statements updated all database implementations updated all domain models created the container for all database Tests

+++ shared/src/commonMain/kotlin/de/darthkali/weefood/datasource/database/WeeFoodDatabaseFactory.kt
@@ -101,8 +101,8 @@ fun List<Recipe_Entity>.toRecipeList(): List<Recipe>{
 fun RecipeIngredient_Entity.toRecipeIngredient(): RecipeIngredient {
     return RecipeIngredient(
         id = id.toInt(),
-        quantity = quantity.toInt(),
-        unit = unit.toInt(),
+        quantity = quantity,
+        unit = unit,
         recipe_id = recipe_id,
         ingredient_id = ingredient_id,
     )
